refactor(chat): split useChat controllers
